Update myScript.js
Added bit more code for a more complete interaction and user experience with the design.
The "backToTop" button is hidden until the window gets to more than 20px from the top.  Added also the condition when the button is clicked, the "active" class to be set to the first li on the nav-list, and remove it from all siblings. For a more smooth experience, I've added a .5s animation to the scrollTop action, that runs on "backToTop" button click.